DRY up upload logic
