+++ app/models/book_upload.rb
@@ -1,43 +1,8 @@
 class BookUpload < ActiveRecord::Base
 
-  require 'csv'
+  include Uploadable
 
-  def save
-    @new_books = Array.new
-    if imported_books.each { |i| make_book(i) }
-      @new_books
-    else
-      imported_books.each_with_index do |product, index|
-        product.errors.full_messages.each do |message|
-          errors.add :base, "Row #{index+2}: #{message}"
-        end
-      end
-      false
-    end
-  end
-
-  def imported_books
-    @imported_books ||= load_imported_books
-  end
-
-  def load_imported_books
-    unassigned_data = open_spreadsheet
-    unassigned_data[0].each { |col| col = col.downcase! }
-    headers = unassigned_data.delete_at(0)
-    books = unassigned_data.map! { |book| Hash[headers.zip(book)] }
-  end
-
-  def open_spreadsheet
-    case File.extname(file.original_filename)
-    when ".csv" then CSV.read(file.path, encoding: "bom|utf-8")
-    when ".txt" then CSV.read(file.path, encoding: "bom|utf-8")
-    # when ".xls" then Excel.new(file.path, nil, :ignore)
-    # when ".xlsx" then Excelx.new(file.path, nil, :ignore)
-    else raise InvalidFileError
-    end
-  end
-
-  def make_book(book_data)
+  def make_object(book_data)
     return false if book_data["title"] == false
 
     if (book = Book.find_by_isbn(book_data["isbn"]))
@@ -53,7 +18,7 @@ class BookUpload < ActiveRecord::Base
                         publication_place:  book_data["publication place"],
                         isbn: book_data["isbn"])
     
-    @new_books << book.id
+    @new_objects << book.id
 
     if book_data["author"]
       make_authors(book_data["author"]).each do |auth| 
@@ -72,9 +37,6 @@ class BookUpload < ActiveRecord::Base
     else
       false
     end
-  end
-
-  class InvalidFileError < RuntimeError
   end
 end
